Reservas Endpoint Compra final reservas y lineas

+++ Models/CaracteristicasSala.cs
@@ -9,13 +9,13 @@ public class CaracteristicasSala
     public int IdCaracteristica { get; set; }
     public string Nombre { get; set; }
     public string Descripcion { get; set; }
-    public double PrecioAniadido { get; set; } = 0; //Ponemos que sea cero porque algunas salas pueden  tener caracteristicas que no tengan precio añadido 
+    public decimal PrecioAniadido { get; set; } = 0; //Ponemos que sea cero porque algunas salas pueden  tener caracteristicas que no tengan precio añadido 
 
 
 
     public CaracteristicasSala() { }
 
-    public CaracteristicasSala(int idCaracteristica, string nombre, string descripcion, double precioAniadido)
+    public CaracteristicasSala(int idCaracteristica, string nombre, string descripcion, decimal precioAniadido)
     {
         IdCaracteristica = idCaracteristica;
         Nombre = nombre;
