+++ Models/TiposPuestosTrabajo.cs
@@ -11,11 +11,11 @@ public class TiposPuestosTrabajo
         public string Nombre { get; set; }
         public string Imagen_URL { get; set; }
         public string Descripcion { get; set; }
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
         
     public TiposPuestosTrabajo() { }
 
-    public TiposPuestosTrabajo(int idTipoPuestoTrabajo, string nombre, string imagen_URL, string descripcion, double precio)
+    public TiposPuestosTrabajo(int idTipoPuestoTrabajo, string nombre, string imagen_URL, string descripcion, decimal precio)
     {
         IdTipoPuestoTrabajo = idTipoPuestoTrabajo;
         Nombre = nombre;
